refactor: restructure repo from scripts/ to src/ ML pipeline layout (#47)
- Replace scripts/dataholder/ with src/strategies/ (Strategy pattern)
- Replace scripts/standalone/ with src/data/ and src/models/
- Extract PostProcessor from HelperMethods god-object
- Add src/cli.py with argparse (replaces custom state-machine parser)
- Add src/strategies/__init__.py with create_strategy() factory
- Add src/data/etl.py with --etl flag for raw data processing
- Add data/input_raw/ with synthetic test data
- Rewrite main.py as minimal pipeline: ETL → Load → Preprocess → Predict → Evaluate → Save
- Move deprecated scripts to archive/, delete scripts/ and tools/
- Delete .vscode/, Makefile
