--- .vscode/settings.json
@@ -1,5 +1,0 @@
-{
-    "python.analysis.extraPaths": [
-        "./scripts"
-    ]
-}
