+++ archive/append_studentcount_and_compute_errors.py
@@ -4,7 +4,7 @@ import numpy as np
 import sys
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "..")))
-from scripts.load_data import load_configuration
+from src.config import load_configuration
 
 
 class AppendStudentCountAndComputeErrors:
