+++ configuration/configuration.json
@@ -11,7 +11,10 @@
         "path_student_count_higher-years": "data/input/student_count_higher-years.xlsx",
         "path_student_volume": "data/input/student_volume.xlsx",
         "path_ratios": "data/input/ratiobestand.xlsx",
-        "path_individual": "data/input/vooraanmeldingen_individueel_DEMO.csv"
+        "path_individual": "data/input/vooraanmeldingen_individueel_DEMO.csv",
+        "path_raw_telbestanden": "data/input_raw/telbestanden",
+        "path_raw_individueel": "data/input_raw/individuele_aanmelddata.csv",
+        "path_raw_afstanden": "data/input_raw/afstanden.xlsx"
     },
     "numerus_fixus": {
         "B Opleiding": 0
