+++ data/input_raw/individuele_aanmelddata.csv
@@ -1,0 +1,6 @@
+Sleutel;Datum Verzoek Inschr;Ingangsdatum;Collegejaar;Datum intrekking vooraanmelding;Inschrijfstatus;Faculteit;Examentype;Croho;Croho groepeernaam;Opleiding;Hoofdopleiding;Eerstejaars croho jaar;Is eerstejaars croho opleiding;Is hogerejaars;BBC ontvangen;Type vooropleiding;Nationaliteit;EER;Geslacht;Geverifieerd adres postcode;Geverifieerd adres plaats;Geverifieerd adres land;Studieadres postcode;Studieadres land;School code eerste vooropleiding;School eerste vooropleiding;Plaats code eerste vooropleiding;Land code eerste vooropleiding;Aantal studenten
+STU00001;15-11-2023;01-09-2024;2024;;Ingeschreven;FdM;Propedeuse Bachelor;56604;B Psychologie;Psychologie;Psychologie;2024;1;0;0;VWO;Nederlandse;J;V;6511AB;Nijmegen;NL;6511AB;NL;30GQ;Stedelijk Gymnasium Nijmegen;NIJM;NL;1
+STU00002;02-12-2023;01-09-2024;2024;;Verzoek tot inschrijving;FdM;Propedeuse Bachelor;56604;B Psychologie;Psychologie;Psychologie;2024;1;0;0;VWO;Duitse;J;M;47608;Geldern;DE;6525HP;NL;DE01;Gymnasium Kleve;KLEV;DE;1
+STU00003;10-01-2024;01-09-2024;2024;;Ingeschreven;FSW;Propedeuse Bachelor;56981;B Politicologie;Politicologie;Politicologie;2024;1;0;0;HAVO;Nederlandse;J;V;5211AB;Den Bosch;NL;5211AB;NL;22KP;Jeroen Bosch College;DENB;NL;1
+STU00004;20-01-2024;01-09-2024;2024;15-03-2024;Geannuleerd;FdR;Master;60139;M Rechtsgeleerdheid;Rechtsgeleerdheid;Rechtsgeleerdheid;2024;1;0;0;WO Bachelor;Nederlandse;J;M;3511AB;Utrecht;NL;6525EC;NL;21PB;Radboud Universiteit;NIJM;NL;1
+STU00005;05-02-2024;01-09-2024;2024;;Verzoek tot inschrijving;FNWI;Propedeuse Bachelor;56857;B Scheikunde;Scheikunde;Scheikunde;2024;1;0;0;VWO;Chinese;N;V;100001;Beijing;CN;6525GD;NL;CN01;Beijing Normal University;BEIJ;CN;1
