+++ data/input_raw/telbestanden/telbestandY2024W01.csv
@@ -1,0 +1,6 @@
+Brincode;Studiejaar;Type_HO;Isatcode;Herkomst;Hogerejaars;Herinschrijving;Aantal;meercode_V
+21PB;2024;B;56604;N;N;N;12;1.29
+21PB;2024;B;56604;E;N;N;3;1.15
+21PB;2024;M;60139;N;N;N;5;1.10
+21PB;2024;B;56981;N;N;N;8;1.22
+21PB;2024;B;56981;R;N;N;2;1.00
