+++ data/input_raw/telbestanden/telbestandY2024W02.csv
@@ -1,0 +1,6 @@
+Brincode;Studiejaar;Type_HO;Isatcode;Herkomst;Hogerejaars;Herinschrijving;Aantal;meercode_V
+21PB;2024;B;56604;N;N;N;18;1.31
+21PB;2024;B;56604;E;N;N;5;1.18
+21PB;2024;M;60139;N;N;N;7;1.12
+21PB;2024;B;56981;N;N;N;14;1.25
+21PB;2024;B;56981;R;N;N;3;1.05
