+++ data/input_raw/telbestanden/telbestandY2024W03.csv
@@ -1,0 +1,6 @@
+Brincode;Studiejaar;Type_HO;Isatcode;Herkomst;Hogerejaars;Herinschrijving;Aantal;meercode_V
+21PB;2024;B;56604;N;N;N;25;1.33
+21PB;2024;B;56604;E;N;N;7;1.20
+21PB;2024;M;60139;N;N;N;9;1.14
+21PB;2024;B;56981;N;N;N;20;1.27
+21PB;2024;B;56981;R;N;N;4;1.08
