+++ src/strategies/base.py
@@ -1,45 +1,36 @@
-from scripts.helper import *
-from scripts.dataholder.helpermethods import *
 from abc import ABC, abstractmethod
 import collections
+import numpy as np
+import pandas as pd
+
+from src.utils.weeks import get_max_week
+from src.output.postprocessor import PostProcessor
 
 
-class Superclass(ABC):
-    """This class is the superclass of the BothDatasets, Cumulative and Individual class. These
-    can be found respectively in botdatasets.py, cumulative.py and individual.py. This class is
-    a subclass of the abc (Abstract Base Classes) module. This module provides the
-    infrastructure for definiing abstract base classes in Python.
-    """
+class PredictionStrategy(ABC):
+    """Base class for prediction strategies (Individual, Cumulative, Combined)."""
 
-    def __init__(self, configuration, helpermethods_initialise_material):
-        # Load in capacity numbers of the numerus fixus studies.
+    def __init__(self, configuration, data_latest, ensemble_weights,
+                 data_studentcount, cwd, data_option, ci_test_n):
         self.numerus_fixus_list = configuration["numerus_fixus"]
 
-        # Load in all the helpermethods used on dataholders as a variable. These methods are
-        # mostly used in main.py for several tasks. Also used to add predicted preregistrations
-        # when calculating the cumulative values.
-        self.helpermethods = HelperMethods(configuration, helpermethods_initialise_material)
+        self.postprocessor = PostProcessor(
+            configuration, data_latest, ensemble_weights,
+            data_studentcount, cwd, data_option, ci_test_n,
+        )
 
-        # programme_filtering and herkomst_filtering will be initialised as empty lists but will
-        # be declared later in set_filtering().
         self.programme_filtering = []
         self.herkomst_filtering = []
         self.examentype_filtering = []
 
-    # Abstract method that must be defined in the subclasses (Indivual, Cumulative, BothDatasets)
-    # itself.
     @abstractmethod
-    def preprocess(self, programme_filtering, herkomst_filtering):
+    def preprocess(self):
         pass
 
-    # Abstract method that must be defined in the subclasses (Indivual, Cumulative, BothDatasets)
-    # itself.
     @abstractmethod
     def predict_nr_of_students(self, predict_year, predict_week, skip_years):
         pass
 
-    # Sets the year and week to predict. These values are specified in the command line when
-    # executing the code. Also looks at the max year and week by looking in the data.
     def set_year_week(self, predict_year, predict_week, data):
         self.predict_year = predict_year
         self.predict_week = predict_week
@@ -47,30 +38,14 @@ class Superclass(ABC):
         self.max_year = data["Collegejaar"].max()
         self.max_week = get_max_week(self.predict_year, self.max_year, data, "Collegejaar")
 
-    # Sets the programme and herkomst filtering. These values are specified in the configuration
-    # file.
     def set_filtering(self, programme_filtering, herkomst_filtering, examentype_filtering):
         self.programme_filtering = programme_filtering
         self.herkomst_filtering = herkomst_filtering
         self.examentype_filtering = examentype_filtering
 
-    # Method that returns a table with information about the data that will be predicted. This will
-    # done repeatedly for every specified week and year to be predicted. This is an example of
-    # what this method would return when we filter on programme 'B Bedrijfskunde' and
-    # 'B Communicatiewetenschap' and herkomst 'NL' and 'EER':
-
-    #          Croho groepeernaam Herkomst  Collegejaar  Weeknummer Examentype Faculteit
-    # 0           B Bedrijfskunde      EER         2024          11   Bachelor       FdM
-    # 1           B Bedrijfskunde       NL         2024          11   Bachelor       FdM
-    # 2  B Communicatiewetenschap      EER         2024          11   Bachelor       FSW
-    # 3  B Communicatiewetenschap       NL         2024          11   Bachelor       FSW
-
-    # This method will only be called when predicting individual values only because this will
-    # done slightly different when predicting cumulative or both datasets.
     def get_data_to_predict(
         self, data, programme_filtering=[], herkomst_filtering=[], examentype_filtering=[]
     ):
-        # These are the columns that will be defined in the table of data to predict.
         predict_dict = {
             "Croho groepeernaam": [],
             "Herkomst": [],
@@ -80,8 +55,6 @@ class Superclass(ABC):
             "Faculteit": [],
         }
 
-        # Take the intersection of the programmes that are in the data and the programmes that
-        # we filter on.
         all_programmes = data["Croho groepeernaam"].unique()
         if programme_filtering != []:
             all_programmes = list(
@@ -90,8 +63,6 @@ class Superclass(ABC):
                 ).elements()
             )
 
-        # Take the intersection of the herkomst that are in the data and the herkomst that we
-        # filter on.
         all_herkomsts = data["Herkomst"].unique()
         if herkomst_filtering != []:
             all_herkomsts = list(
@@ -109,8 +80,6 @@ class Superclass(ABC):
                 ).elements()
             )
 
-        # Add one line in the table per unique programme, examentype and herkomst that has to
-        # be predicted.
         for programme in np.sort(all_programmes):
             available_examentypes_for_programme = data[
                 (data["Croho groepeernaam"] == programme)
